add message utility class
